Add special handling of $PATH variable
This is nessicary because in xonsh $PATH is a list (as it should be),
while in POSIX-compliant shells it's a string

This adds nessicary logic to the runtime to translate back and forth.
It comes with some limitations (can only add to beginning or end &
cannot remove), but 99% of the time that should work.

As always, if any problems arise zsh2xonsh will throw an exception
with a reasonable error message :)

This is an issue for translating $(brew shellenv), because
that modifies the $PATH directly

However my environment files do this less often because they use
`append_path` commands to modify things indirectly :)