Add an option to import the runtime
This is on by default, and adds `from zsh2xonsh import runtime`
to the top of the generated output.

The previous behavior was to blindly assume that a `runtime` was already
loaded (thorwing NameError if missing)

There is an option `--no-runtime` to disable it if you know it's
already imported (to get the old behavior back).